+++ DonutCoin/client/src/Components/Mypage/MypageForm.js
@@ -1,11 +1,9 @@
-import React, { useRef, useState } from 'react';
-import { withRouter } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
-import { registerUser, checkUser } from '../../reducer/action/user';
-
-
-import { useForm } from 'react-hook-form';
+import React, { useEffect, useRef, useState } from 'react';
 import styled from "styled-components";
+import { Link } from 'react-router-dom';
+import { auth } from '../../reducer/action/user';
+import { useSelector, useDispatch } from 'react-redux';
+import axios from 'axios';
 
 
 const St = {
@@ -17,9 +15,9 @@ const St = {
     justify-content: center;
     background-color: rgb(231, 234, 239);
   `,
-  Login: styled.form`
+  Mypage: styled.form`
     width: 30%;
-    height: 600px;
+    height: 700px;
     background: white;
     border-radius: 20px;
     display: flex;
@@ -34,11 +32,15 @@ const St = {
     font-weight: bolder;
     text-shadow: 1px 2px 3px black;
   `,
-  Login_sns: styled.div`
-    padding: 20px;
-    display: flex;
+  Id: styled.div`
+    margin-top: 20px;
+    width: 80%;
   `,
-  Login_id: styled.div`
+  Pw: styled.div`
+    margin-top: 20px;
+    width: 80%;
+  `,
+  Name: styled.div`
     margin-top: 20px;
     width: 80%;
   `,
@@ -50,15 +52,7 @@ const St = {
     padding: 0px 20px;
     border: 1px solid lightgray;
     outline: none;
-  `,
-  Login_pw: styled.div`
-    margin-top: 20px;
-    width: 80%;
-  `,
-  Join_name: styled.div`
-    margin-top: 20px;
-    width: 80%;
-`,
+    `,
   Submit: styled.div`
     margin-top: 50px;
     width: 80%;
@@ -79,130 +73,54 @@ const St = {
         color: white;
     }
     `,
-  ToJoin: styled.a`
-    margin: auto;
-    text-decoration-line: none;
-    text-decoration-line: none;
-    `,
-
 };
-function MypageForm(props) {
-  const {
-    register,
-    handleSubmit,
-    watch,
-    setError,
-    formState: { errors },
-  } = useForm({
-    mode: 'onTouched',
-  });
+const MypageForm = (props, { isRootURL }) => {
   const dispatch = useDispatch();
-  const password = useRef();
-  password.current = watch('password');
 
-  const [ShowPassword, setShowPassword] = useState(false);
-  const handleVisibility = () => {
-    setShowPassword(!ShowPassword);
+  const [userEmail, setUserEmail] = useState('');
+  const [userName, setUserName] = useState('');
+  const [userPassword, setUserPassword] = useState('');
+
+  const user = useSelector(state => state.user);
+
+  const getName = () => {
+    dispatch(auth()).then(response => {
+      if (response.payload.userData != null) {
+        setUserName(response.payload.userData.name);
+      }
+    });
   };
 
-  const onSubmit = async data => {
-    // console.log(data);
-    try {
-      await dispatch(checkUser(data.email))
-        .then(response => {
-          console.log(response)
-          if (response.payload.success) {
-            dispatch(registerUser(data));
-            alert(`${data.name}님 회원가입을 축하드립니다.`);
-            props.history.push('/login');
-          } else {
-            setError('email', {
-              type: 'validate',
-              message: response.payload.message,
-            }
-            );
-          }
-        })
-        .catch(error => {
-          console.log('response: ', error.response);
-        });
-    } catch (err) {
-      console.log(err);
-    }
-  };
+  useEffect(() => {
+    getName();
+  }, []);
 
   return (
     <St.Container>
-      <St.Login onSubmit={handleSubmit(onSubmit)}>
+      <St.Mypage>
         <St.Head>MY PAGE</St.Head>
-        <St.Login_id>
+        <St.Id>
           <h4>E-mail</h4>
-          <St.Input
-            id="email"
-            name="email"
-            type="email"
-            {...register('email', {
-              required: '이메일을 입력해주세요.',
-              pattern: {
-                // value: /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i,
-                message: '이메일 형식이 올바르지 않습니다.',
-              },
-            })} />
-        </St.Login_id>
-        <St.Join_name>
+          <St.Input/>
+        </St.Id>
+        <St.Name>
           <h4>Name</h4>
-          <St.Input
-            id="name"
-            name="name"
-            type="text"
-            {...register('name', {
-              required: true,
-              minLength: true,
-              minLength: 2,
-              maxLength: 8,
-            })} />
-        </St.Join_name>
-        <St.Login_pw>
+          <St.Input/>
+        </St.Name>
+        <St.Pw>
           <h4>Password</h4>
-          <St.Input id="password"
-            name="password"
-            type={ShowPassword ? 'text' : 'password'}
-            {...register('password', {
-              required: true,
-              minLength: 8,
-              maxLength: 20,
-              validate: {
-                checkLang: value =>
-                  ![/[ㄱ-ㅎ|ㅏ-ㅣ|가-힣]/].every(pattern =>
-                    pattern.test(value),
-                  ),
-                // checkLower: value =>
-                //   [/[a-z]/].every(pattern => pattern.test(value)),
-                // checkUpper: value =>
-                //   [/[A-Z]/].every(pattern => pattern.test(value)),
-                // checkNumber: value =>
-                //   [/[0-9]/].every(pattern => pattern.test(value)),
-                // checkSpec: value =>
-                //   [/[^a-zA-Z0-9]/].every(pattern => pattern.test(value)),
-              },
-            })} />
-        </St.Login_pw>
-        <St.Login_pw>
+          <St.Input />
+        </St.Pw>
+        <St.Pw>
           <h4>My Wallet Address</h4>
-          <St.Input id="confirmpassword"
-            name="confirmpassword"
-            type="password"
-            {...register('passwordConfirm', {
-              required: true,
-              validate: value => value === password.current,
-            })} />
-        </St.Login_pw>
+          <St.Input/>
+        </St.Pw>
         <St.Submit>
-          <St.Submit_button type="submit" onClick={handleSubmit(onSubmit)}>Edit</St.Submit_button>
+          <St.Submit_button>Edit</St.Submit_button>
         </St.Submit>
-      </St.Login>
+      </St.Mypage>
     </St.Container>
   )
-}
+};
 
-export default withRouter(MypageForm);
+export default MypageForm;
